Split API handlers into separate go files
