+++ api/filters.go
@@ -6,8 +6,6 @@ import (
 	"errors"
 	"net/http"
 	"regexp"
-	"time"
-
 	"github.com/ONSdigital/dp-filter-api/models"
 	"github.com/ONSdigital/go-ns/log"
 	"github.com/gorilla/mux"
@@ -193,383 +191,6 @@ func (api *FilterAPI) getFilterBlueprint(w http.ResponseWriter, r *http.Request)
 	log.Info("got filter blueprint", logData)
 }
 
-func (api *FilterAPI) getFilterBlueprintDimensions(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	logData := log.Data{"filter_blueprint_id": filterID}
-	log.Info("getting filter blueprint dimensions", logData)
-
-	filter, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.ErrorC("unable to get dimensions for filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	logData["dimensions"] = filter.Dimensions
-
-	if len(filter.Dimensions) == 0 {
-		log.Error(errDimensionNotFound, log.Data{"filter_blueprint_id": filterID})
-		setErrorCode(w, errDimensionNotFound)
-		return
-	}
-
-	bytes, err := json.Marshal(filter.Dimensions)
-	if err != nil {
-		log.ErrorC("failed to marshal filter blueprint dimensions into bytes", err, logData)
-		http.Error(w, internalError, http.StatusInternalServerError)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-	_, err = w.Write(bytes)
-	if err != nil {
-		log.ErrorC("failed to write bytes for http response", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	log.Info("got dimensions for filter blueprint", logData)
-}
-
-func (api *FilterAPI) getFilterBlueprintDimension(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	logData := log.Data{
-		"filter_blueprint_id": filterID,
-		"dimension":           name,
-	}
-	log.Info("getting filter blueprint dimension", logData)
-
-	if _, err := api.getFilter(r.Context(), filterID); err != nil {
-		log.Error(err, log.Data{"filter_blueprint_id": filterID})
-		setErrorCode(w, err)
-		return
-	}
-
-	if err := api.dataStore.GetFilterDimension(filterID, name); err != nil {
-		log.Error(err, log.Data{"filter_blueprint_id": filterID, "dimension": name})
-		setErrorCode(w, err)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusNoContent)
-
-	log.Info("got filtered blueprint dimension", logData)
-}
-
-func (api *FilterAPI) removeFilterBlueprintDimension(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	logData := log.Data{
-		"filter_blueprint_id": filterID,
-		"dimension":           name,
-	}
-	log.Info("removing filter blueprint dimension", logData)
-
-	filter, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.Error(err, log.Data{"filter_blueprint_id": filterID})
-		setErrorCode(w, err)
-		return
-	}
-
-	if filter.State == models.SubmittedState {
-		log.Error(errForbidden, log.Data{"filter_blueprint_id": filterID})
-		setErrorCode(w, errForbidden)
-		return
-	}
-
-	if err := api.dataStore.RemoveFilterDimension(filterID, name); err != nil {
-		log.ErrorC("unable to remove dimension from filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-
-	log.Info("delete dimension from filter blueprint", logData)
-}
-
-func (api *FilterAPI) addFilterBlueprintDimension(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	logData := log.Data{
-		"filter_blueprint_id": filterID,
-		"dimension":           name,
-	}
-	log.Info("add filter blueprint dimension", logData)
-
-	options, err := models.CreateDimensionOptions(r.Body)
-	if err != nil {
-		log.ErrorC("unable to unmarshal request body", err, logData)
-		http.Error(w, badRequest, http.StatusBadRequest)
-		return
-	}
-
-	filterBlueprint, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.Error(err, log.Data{"filter_blueprint_id": filterID})
-		setErrorCode(w, err)
-		return
-	}
-	logData["current_filter_blueprint"] = filterBlueprint
-
-	if filterBlueprint.State == models.SubmittedState {
-		log.Error(errForbidden, log.Data{"filter_blueprint_id": filterID})
-		setErrorCode(w, errForbidden)
-		return
-	}
-
-	if err = api.checkNewFilterDimension(r.Context(), name, options, *filterBlueprint.Dataset); err != nil {
-		log.ErrorC("unable to get filter blueprint", err, logData)
-		if err == ErrVersionNotFound {
-			setErrorCode(w, err, statusUnprocessableEntity)
-			return
-		}
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	if err = api.dataStore.AddFilterDimension(filterID, name, options, filterBlueprint.Dimensions); err != nil {
-		log.ErrorC("failed to add dimension to filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusCreated)
-
-	log.Info("created new dimension for filter blueprint", logData)
-}
-
-func (api *FilterAPI) getFilterBlueprintDimensionOptions(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	logData := log.Data{
-		"filter_blueprint_id": filterID,
-		"dimension":           name,
-	}
-	log.Info("get filter blueprint dimension options", logData)
-
-	filter, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.ErrorC("unable to get dimension options for filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	var options []models.DimensionOption
-	dimensionFound := false
-	for _, dimension := range filter.Dimensions {
-
-		if dimension.Name == name {
-			dimensionFound = true
-			for _, option := range dimension.Options {
-				url := fmt.Sprintf("%s/filter/%s/dimensions/%s/option/%s", api.host, filterID, dimension.Name, option)
-				dimensionOption := models.DimensionOption{Option: option, DimensionOptionURL: url}
-				options = append(options, dimensionOption)
-			}
-
-		}
-	}
-
-	if !dimensionFound {
-		log.Error(errDimensionNotFound, log.Data{"filter_blueprint_id": filterID, "dimension_name": name})
-		setErrorCode(w, errDimensionNotFound)
-		return
-	}
-
-	logData["options"] = options
-
-	bytes, err := json.Marshal(options)
-	if err != nil {
-		log.ErrorC("failed to marshal filter blueprint dimension options into bytes", err, logData)
-		http.Error(w, internalError, http.StatusInternalServerError)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-	_, err = w.Write(bytes)
-	if err != nil {
-		log.ErrorC("failed to write bytes for http response", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	log.Info("got dimension options for filter blueprint", logData)
-}
-
-func (api *FilterAPI) getFilterBlueprintDimensionOption(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	option := vars["option"]
-	logData := log.Data{
-		"filter_blueprint_id": filterID,
-		"dimension":           name,
-		"option":              option,
-	}
-	log.Info("get filter blueprint dimension option", logData)
-
-	filter, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.ErrorC("unable to get dimension option for filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	dimensionFound := false
-	optionFound := false
-	for _, d := range filter.Dimensions {
-		if d.Name == name {
-			dimensionFound = true
-			for _, o := range d.Options {
-				if o == option {
-					optionFound = true
-				}
-			}
-		}
-	}
-
-	if !dimensionFound {
-		log.Error(errDimensionNotFound, logData)
-		setErrorCode(w, errDimensionNotFound)
-		return
-	}
-
-	if !optionFound {
-		log.Error(errOptionNotFound, logData)
-		setErrorCode(w, errOptionNotFound)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusNoContent)
-
-	log.Info("got dimension option for filter blueprint", logData)
-}
-
-func (api *FilterAPI) addFilterBlueprintDimensionOption(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	option := vars["option"]
-	logData := log.Data{"filter_id": filterID, "dimension_name": name, "dimension_option": option}
-
-	filterBlueprint, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.Error(err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	if filterBlueprint.State == models.SubmittedState {
-		log.Error(errForbidden, logData)
-		setErrorCode(w, errForbidden, filterBlueprint.State)
-		return
-	}
-
-	// FIXME - Once dataset API has an endpoint to check single option exists,
-	// refactor code below instead of creating an AddDimension object from the
-	// AddDimensionOption object (to be able to use checkNewFilterDimension method)
-	if err = api.checkNewFilterDimension(r.Context(), name, []string{option}, *filterBlueprint.Dataset); err != nil {
-		if err == ErrVersionNotFound {
-			log.ErrorC("failed to select valid version", err, logData)
-			setErrorCode(w, err, statusUnprocessableEntity)
-			return
-		}
-
-		if incorrectDimensionOptions.MatchString(err.Error()) {
-			log.ErrorC("failed to select valid filter dimension options", err, logData)
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		if incorrectDimension.MatchString(err.Error()) {
-			log.ErrorC("failed to select valid filter dimension", err, logData)
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		log.ErrorC("failed to successfully check filter dimensions", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	if err := api.dataStore.AddFilterDimensionOption(filterID, name, option); err != nil {
-		log.ErrorC("failed to add dimension option to filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusCreated)
-
-	log.Info("created new dimension option for filter blueprint", logData)
-}
-
-func (api *FilterAPI) removeFilterBlueprintDimensionOption(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_blueprint_id"]
-	name := vars["name"]
-	option := vars["option"]
-
-	logData := log.Data{
-		"filter_blueprint_id": filterID,
-		"dimension":           name,
-		"option":              option,
-	}
-	log.Info("remove filter blueprint dimension option", logData)
-
-	filterBlueprint, err := api.getFilter(r.Context(), filterID)
-	if err != nil {
-		log.Error(err, log.Data{"filter_id": filterID})
-		setErrorCode(w, err)
-		return
-	}
-
-	// Check if dimension exists
-	var hasDimension bool
-	for _, dimension := range filterBlueprint.Dimensions {
-		if dimension.Name == name {
-			hasDimension = true
-			break
-		}
-	}
-
-	if !hasDimension {
-		log.Error(errDimensionNotFound, logData)
-		setErrorCode(w, errDimensionNotFound)
-		return
-	}
-
-	if filterBlueprint.State == models.SubmittedState {
-		log.Error(errForbidden, logData)
-		setErrorCode(w, errForbidden)
-		return
-	}
-
-	if err = api.dataStore.RemoveFilterDimensionOption(filterID, name, option); err != nil {
-		log.ErrorC("unable to remove dimension option from filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-
-	log.Info("delete dimension option on filter blueprint", logData)
-}
-
 func (api *FilterAPI) updateFilterBlueprint(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	filterID := vars["filter_blueprint_id"]
@@ -682,221 +303,6 @@ func (api *FilterAPI) updateFilterBlueprint(w http.ResponseWriter, r *http.Reque
 	log.Info("filter blueprint updated", logData)
 }
 
-func createNewFilter(filter *models.Filter, currentFilter *models.Filter) (newFilter *models.Filter, versionHasChanged bool) {
-	newFilter = currentFilter
-
-	if filter.Dataset != nil {
-		if filter.Dataset.Version != 0 && filter.Dataset.Version != currentFilter.Dataset.Version {
-			versionHasChanged = true
-			newFilter.Dataset.Version = filter.Dataset.Version
-		}
-
-		if &filter.Events != nil {
-			if filter.Events.Info != nil {
-				newFilter.Events.Info = append(newFilter.Events.Info, filter.Events.Info...)
-			}
-
-			if filter.Events.Error != nil {
-				newFilter.Events.Error = append(newFilter.Events.Error, filter.Events.Error...)
-			}
-		}
-	}
-
-	return
-}
-
-func (api *FilterAPI) getFilterOutput(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterOutputID := vars["filter_output_id"]
-
-	logData := log.Data{"filter_output_id": filterOutputID}
-	log.Info("getting filter output", logData)
-
-	filterOutput, err := api.getOutput(r, filterOutputID)
-	if err != nil {
-		log.ErrorC("unable to get filter output", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-	logData["filter_output"] = filterOutput
-
-	bytes, err := json.Marshal(filterOutput)
-	if err != nil {
-		log.ErrorC("failed to marshal filter output into bytes", err, logData)
-		http.Error(w, internalError, http.StatusInternalServerError)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-	_, err = w.Write(bytes)
-	if err != nil {
-		log.ErrorC("failed to write bytes for http response", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	log.Info("got filter output", logData)
-}
-
-func (api *FilterAPI) updateFilterOutput(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterOutputID := vars["filter_output_id"]
-
-	logData := log.Data{"filter_output_id": filterOutputID}
-	log.Info("updating filter output", logData)
-
-	if !common.IsCallerPresent(r.Context()) {
-		err := errors.New("Not authorised")
-		log.ErrorC("failed to update filter output", err, logData)
-		setErrorCode(w, errNoAuthHeader)
-		return
-	}
-
-	filterOutput, err := models.CreateFilter(r.Body)
-	if err != nil {
-		log.ErrorC("unable to unmarshal request body", err, logData)
-		http.Error(w, badRequest, http.StatusBadRequest)
-		return
-	}
-	logData["filter_output"] = filterOutput
-
-	// check filter output resource for current downloads and published flag
-	previousFilterOutput, err := api.dataStore.GetFilterOutput(filterOutputID)
-	if err != nil {
-		log.ErrorC("unable to get current filter output", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	if err = filterOutput.ValidateFilterOutputUpdate(previousFilterOutput); err != nil {
-		log.ErrorC("filter output failed validation", err, logData)
-		http.Error(w, err.Error(), http.StatusForbidden)
-		return
-	}
-
-	filterOutput.FilterID = filterOutputID
-
-	// Set the published flag to the value currently stored on filter output resources
-	// unless the request contains an update to the flag
-	if previousFilterOutput.Published != nil && *previousFilterOutput.Published == models.Published {
-		filterOutput.Published = &models.Published
-	}
-
-	filterOutputUpdate := buildDownloadsObject(previousFilterOutput, filterOutput, api.downloadServiceURL)
-
-	if err = api.dataStore.UpdateFilterOutput(filterOutputUpdate); err != nil {
-		log.ErrorC("unable to update filter blueprint", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-
-	log.Info("update filter output", logData)
-}
-
-func (api *FilterAPI) createFilterOutputResource(newFilter *models.Filter, filterBlueprintID string) (models.Filter, error) {
-	filterOutput := *newFilter
-	filterOutput.FilterID = uuid.NewV4().String()
-	filterOutput.State = models.CreatedState
-	filterOutput.Links.Self.HRef = fmt.Sprintf("%s/filter-outputs/%s", api.host, filterOutput.FilterID)
-	filterOutput.Links.Dimensions.HRef = ""
-	filterOutput.Links.FilterBlueprint.HRef = fmt.Sprintf("%s/filters/%s", api.host, filterBlueprintID)
-	filterOutput.Links.FilterBlueprint.ID = filterBlueprintID
-	filterOutput.LastUpdated = time.Now()
-
-	// Clear out any event information to output document
-	filterOutput.Events = models.Events{}
-
-	// Downloads object should exist for filter output resource
-	// even if it they are empty
-	filterOutput.Downloads = &models.Downloads{
-		CSV: &models.DownloadItem{},
-		XLS: &models.DownloadItem{},
-	}
-
-	// Remove dimension url from output filter resource
-	for i := range newFilter.Dimensions {
-		filterOutput.Dimensions[i].URL = ""
-	}
-
-	if newFilter.Published == &models.Published {
-		filterOutput.Published = &models.Published
-	}
-
-	if err := api.dataStore.CreateFilterOutput(&filterOutput); err != nil {
-		log.ErrorC("unable to create filter output", err, log.Data{"filter_output": filterOutput})
-		return models.Filter{}, err
-	}
-
-	return filterOutput, api.outputQueue.Queue(&filterOutput)
-}
-
-func (api *FilterAPI) getFilterOutputPreview(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	filterID := vars["filter_output_id"]
-	requestedLimit := r.URL.Query().Get("limit")
-
-	logData := log.Data{"filter_output_id": filterID}
-	log.Info("get filter output preview", logData)
-
-	var limit = 20
-	var err error
-
-	if requestedLimit != "" {
-		limit, err = strconv.Atoi(requestedLimit)
-		if err != nil {
-			logData["requested_limit"] = requestedLimit
-			log.ErrorC("requested limit is not a number", err, logData)
-			http.Error(w, errRequestLimitNotNumber.Error(), http.StatusBadRequest)
-			return
-		}
-	}
-	logData["limit"] = limit
-
-	filterOutput, err := api.getOutput(r, filterID)
-	if err != nil {
-		log.ErrorC("failed to find filter output", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	logData["filter_output_dimensions"] = filterOutput.Dimensions
-
-	if len(filterOutput.Dimensions) == 0 {
-		log.ErrorC("no dimensions are present in the filter", errMissingDimensions, logData)
-		http.Error(w, "no dimensions are present in the filter", http.StatusBadRequest)
-		return
-	}
-
-	data, err := api.preview.GetPreview(filterOutput, limit)
-	if err != nil {
-		log.ErrorC("failed to query the graph database", err, logData)
-		http.Error(w, "internal server error", http.StatusInternalServerError)
-		return
-	}
-
-	bytes, err := json.Marshal(data)
-	if err != nil {
-		log.ErrorC("failed to marshal preview of filter ouput into bytes", err, logData)
-		http.Error(w, internalError, http.StatusInternalServerError)
-		return
-	}
-
-	setJSONContentType(w)
-	w.WriteHeader(http.StatusOK)
-	_, err = w.Write(bytes)
-	if err != nil {
-		log.ErrorC("failed to write bytes for http response", err, logData)
-		setErrorCode(w, err)
-		return
-	}
-
-	log.Info("preview filter output", logData)
-}
-
 func (api *FilterAPI) getFilter(ctx context.Context, filterID string) (*models.Filter, error) {
 	filter, err := api.dataStore.GetFilter(filterID)
 	if err != nil {
@@ -929,63 +335,6 @@ func (api *FilterAPI) getFilter(ctx context.Context, filterID string) (*models.F
 	}
 
 	return nil, errNotFound
-}
-
-func (api *FilterAPI) getOutput(r *http.Request, filterID string) (*models.Filter, error) {
-
-	ctx := r.Context()
-	logData := log.Data{"filter_output_id": filterID}
-
-	output, err := api.dataStore.GetFilterOutput(filterID)
-	if err != nil {
-		log.Error(err, logData)
-		return nil, err
-	}
-
-	logData["filter_blueprint_id"] = output.Links.FilterBlueprint.ID
-
-	// Hide private download links if request is not authenticated
-	if r.Header.Get(common.DownloadServiceHeaderKey) != api.downloadServiceToken {
-
-		log.Info("a valid download service token has not been provided. hiding private links", logData)
-
-		if output.Downloads != nil {
-			if output.Downloads.CSV != nil {
-				output.Downloads.CSV.Private = ""
-			}
-			if output.Downloads.XLS != nil {
-				output.Downloads.XLS.Private = ""
-			}
-		}
-	} else {
-		log.Info("a valid download service token has been provided. not hiding private links", logData)
-	}
-
-	//only return the filter if it is for published data or via authenticated request
-	if output.Published != nil && *output.Published == models.Published || common.IsCallerPresent(ctx) {
-		return output, nil
-	}
-
-	log.Info("unauthenticated request to access unpublished filter output", logData)
-
-	filter, err := api.getFilter(ctx, output.Links.FilterBlueprint.ID)
-	if err != nil {
-		log.Error(errors.New("failed to retrieve filter blueprint"), logData)
-		return nil, errFilterOutputNotFound
-	}
-
-	//filter has been published since output was last requested, so update output and return
-	if filter.Published != nil && *filter.Published == models.Published {
-		output.Published = &models.Published
-		if err := api.dataStore.UpdateFilterOutput(output); err != nil {
-			log.Error(err, logData)
-			return nil, errFilterOutputNotFound
-		}
-
-		return output, nil
-	}
-
-	return nil, errFilterOutputNotFound
 }
 
 func (api *FilterAPI) checkFilterOptions(ctx context.Context, newFilter *models.Filter, version *models.Version) error {
@@ -1040,107 +389,31 @@ func (api *FilterAPI) checkFilterOptions(ctx context.Context, newFilter *models.
 	return nil
 }
 
-func (api *FilterAPI) checkNewFilterDimension(ctx context.Context, name string, options []string, dataset models.Dataset) error {
-	logData := log.Data{"dimension_name": name, "dimension_options": options, "dataset": dataset}
-	log.Info("check filter dimensions and dimension options before calling api, see version number", logData)
-
-	// FIXME - We should be calling dimension endpoint on dataset API to check if
-	// dimension exists but this endpoint doesn't exist yet so call dimension
-	// list endpoint and iterate over items to find if dimension exists
-	datasetDimensions, err := api.datasetAPI.GetVersionDimensions(ctx, dataset)
-	if err != nil {
-		log.ErrorC("failed to retreive a list of dimensions from the dataset API", err, logData)
-		return err
-	}
-
-	dimension := models.Dimension{
-		Name:    name,
-		Options: options,
-	}
-
-	if err = models.ValidateFilterDimensions([]models.Dimension{dimension}, datasetDimensions); err != nil {
-		log.ErrorC("filter dimensions failed validation", err, logData)
-		return err
-	}
-
-	// Call dimension options endpoint
-	datasetDimensionOptions, err := api.datasetAPI.GetVersionDimensionOptions(ctx, dataset, dimension.Name)
-	if err != nil {
-		log.ErrorC("failed to retreive a list of dimension options from the dataset API", err, logData)
-		return err
-	}
-
-	var incorrectDimensionOptions []string
-	incorrectOptions := models.ValidateFilterDimensionOptions(dimension.Options, datasetDimensionOptions)
-	if incorrectOptions != nil {
-		incorrectDimensionOptions = append(incorrectDimensionOptions, incorrectOptions...)
-	}
-
-	if incorrectDimensionOptions != nil {
-		err = fmt.Errorf("Bad request - incorrect dimension options chosen: %v", incorrectDimensionOptions)
-		log.ErrorC("incorrect dimension options chosen", err, logData)
-		return err
-	}
-
-	return nil
-}
-
-func buildDownloadsObject(previousFilterOutput, filterOutput *models.Filter, downloadServiceURL string) *models.Filter {
-
-	if filterOutput.Downloads == nil {
-		filterOutput.Downloads = previousFilterOutput.Downloads
-		return filterOutput
-	}
-
-	if filterOutput.Downloads.CSV != nil {
-
-		filterOutput.Downloads.CSV.HRef = downloadServiceURL + "/downloads/filter-outputs/" + previousFilterOutput.FilterID + ".csv"
-
-		if previousFilterOutput.Downloads != nil && previousFilterOutput.Downloads.CSV != nil {
-
-			if filterOutput.Downloads.CSV.Size == "" {
-				filterOutput.Downloads.CSV.Size = previousFilterOutput.Downloads.CSV.Size
-			}
-			if filterOutput.Downloads.CSV.Private == "" {
-				filterOutput.Downloads.CSV.Private = previousFilterOutput.Downloads.CSV.Private
-			}
-			if filterOutput.Downloads.CSV.Public == "" {
-				filterOutput.Downloads.CSV.Public = previousFilterOutput.Downloads.CSV.Public
-			}
-		}
-	} else {
-		if previousFilterOutput.Downloads != nil {
-			filterOutput.Downloads.CSV = previousFilterOutput.Downloads.CSV
-		}
-	}
-
-	if filterOutput.Downloads.XLS != nil {
-
-		filterOutput.Downloads.XLS.HRef = downloadServiceURL + "/downloads/filter-outputs/" + previousFilterOutput.FilterID + ".xlsx"
-
-		if previousFilterOutput.Downloads != nil && previousFilterOutput.Downloads.XLS != nil {
-
-			if filterOutput.Downloads.XLS.Size == "" {
-				filterOutput.Downloads.XLS.Size = previousFilterOutput.Downloads.XLS.Size
-			}
-			if filterOutput.Downloads.XLS.Private == "" {
-				filterOutput.Downloads.XLS.Private = previousFilterOutput.Downloads.XLS.Private
-			}
-			if filterOutput.Downloads.XLS.Public == "" {
-				filterOutput.Downloads.XLS.Public = previousFilterOutput.Downloads.XLS.Public
-			}
-		}
-	} else {
-		if previousFilterOutput.Downloads != nil {
-			filterOutput.Downloads.XLS = previousFilterOutput.Downloads.XLS
-		}
-	}
-
-	return filterOutput
-}
-
 func setJSONContentType(w http.ResponseWriter) {
 	w.Header().Set("Content-Type", "application/json")
+}
+
+func createNewFilter(filter *models.Filter, currentFilter *models.Filter) (newFilter *models.Filter, versionHasChanged bool) {
+	newFilter = currentFilter
+
+	if filter.Dataset != nil {
+		if filter.Dataset.Version != 0 && filter.Dataset.Version != currentFilter.Dataset.Version {
+			versionHasChanged = true
+			newFilter.Dataset.Version = filter.Dataset.Version
+		}
+
+		if &filter.Events != nil {
+			if filter.Events.Info != nil {
+				newFilter.Events.Info = append(newFilter.Events.Info, filter.Events.Info...)
+			}
+
+			if filter.Events.Error != nil {
+				newFilter.Events.Error = append(newFilter.Events.Error, filter.Events.Error...)
+			}
+		}
+	}
+
+	return
 }
 
 func setErrorCode(w http.ResponseWriter, err error, typ ...string) {
